restructured the route
